+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.4.4
+
+* Fixes a timer bug.
+
+## 0.4.0
+
+* Add a custom row builder
+
+## 0.3.1
+
+* Add option to clear selected rows
+
+## 0.3.0
+
+* Add refresh listener
+* Add separated widget to build custom filters.
+
+## 0.2.0
+
+* Add widget to build custom filters.
+
 ## 0.1.0
 
 * Initial release.